+++ collectors/bzimport/convertors.py
@@ -974,8 +974,6 @@ class FlawConvertor(BugzillaGroupsConvertorMixin):
                     )
                 )
 
-        flaw.is_major_incident = False
-
         for flag in self.flags:
             flag_name = flag["name"]
             if flag_name not in self.FLAGS_META:
@@ -988,12 +986,6 @@ class FlawConvertor(BugzillaGroupsConvertorMixin):
                     [flag],
                 )
             )
-
-            if flag_name in ["hightouch", "hightouch-lite"] and flag["status"] in [
-                "?",
-                "+",
-            ]:
-                flaw.is_major_incident = True
 
         return meta